+++ sub_controller/OpenCM9.04_Internal_USB/Projects/Virtual_COM_Port/src/usb_endp.c
@@ -99,7 +99,7 @@ void EP1_IN_Callback (void)
 *******************************************************************************/
 void EP3_OUT_Callback(void)
 {
-  uint16_t USB_Rx_Cnt;
+  uint32_t USB_Rx_Cnt;
   
   /* Get the received data buffer and update the counter */
   USB_Rx_Cnt = USB_SIL_Read(EP3_OUT, USB_Rx_Buffer);
